+++ common/src/main/kotlin/com/nbp/cobblemon_kubejs/kubejs/client/CobblemonClientJSBinding.kt
@@ -1,0 +1,6 @@
+package com.nbp.cobblemon_kubejs.kubejs.client
+
+object CobblemonClientJSBinding {
+    @JvmField
+    val pokedex = ClientPokedexJSApi
+}